requirements.txt
farm-haystack==0.4.0
sumeval==0.2.2